gcp/requirements.txt
tensorflow==2.9.1
google-cloud-storage==2.4.0
Pillow==8.1.0